adds seed data with hands and solutions to mongodb
