+++ data/db/WiredTiger
@@ -1,0 +1,2 @@
+WiredTiger
+WiredTiger 2.8.1: (March 24, 2016)
