+++ package.json
@@ -11,8 +11,8 @@
   "dependencies": {
     "express": "4.x",
     "express-handlebars": "1.1.x",
-    "mongoose": "^4.2.5",
     "fast-csv": "2.3.0",
+    "mongoose": "^4.7.2",
     "socket.io": "1.1.x"
   },
   "engines": {
