+++ static/etc/hands.csv
@@ -1,4 +1,4 @@
-hand,possiblesolution
+cards,solution
 1 1 1 8,(1+1+1)×8
 1 1 1 11,(1+1)×(1+11)
 1 1 1 12,(1+1)×1×12
